+++ sdk/servicebus/service-bus/src/serializers/subscriptionResourceSerializer.ts
@@ -91,8 +91,7 @@ export function buildSubscription(rawSubscription: any): SubscriptionDescription
     forwardTo: getStringOrUndefined(rawSubscription[Constants.FORWARD_TO]),
     userMetadata: rawSubscription[Constants.USER_METADATA],
 
-    status: getString(rawSubscription[Constants.STATUS], "status") as EntityStatus,
-    messageCount: getInteger(rawSubscription[Constants.MESSAGE_COUNT], "messageCount")
+    status: getString(rawSubscription[Constants.STATUS], "status") as EntityStatus
   };
 }
 
@@ -218,9 +217,6 @@ export interface SubscriptionDescription {
    * such as "PT1M" for 1 minute, "PT5S" for 5 seconds.
    */
   autoDeleteOnIdle?: string;
-
-  // TODO: will be removed once the RuntimeInfo APIs are added
-  messageCount?: number;
 }
 
 /**
@@ -321,8 +317,6 @@ export interface InternalSubscriptionOptions {
 }
 
 /**
- * @internal
- * @ignore
  * Represents runtime info attributes of a subscription entity
  */
 export interface SubscriptionRuntimeInfo {