+++ sdk/servicebus/service-bus/src/serviceBusAtomManagementClient.ts
@@ -1,56 +1,63 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { parseConnectionString, SharedKeyCredential } from "@azure/core-amqp";
 import {
   HttpOperationResponse,
+  proxyPolicy,
   ProxySettings,
   RequestPolicyFactory,
   RestError,
   ServiceClient,
   ServiceClientOptions,
-  URLBuilder,
-  WebResource,
-  proxyPolicy,
   signingPolicy,
   stripRequest,
-  stripResponse
+  stripResponse,
+  URLBuilder,
+  WebResource
 } from "@azure/core-http";
-
-import { parseConnectionString, SharedKeyCredential, TokenCredential } from "@azure/core-amqp";
-
-import { AtomXmlSerializer, executeAtomXmlOperation } from "./util/atomXmlHelper";
-
 import * as log from "./log";
-import { SasServiceClientCredentials } from "./util/sasServiceClientCredentials";
-import * as Constants from "./util/constants";
-
 import {
+  buildNamespace,
+  NamespaceProperties,
+  NamespaceResourceSerializer
+} from "./serializers/namespaceResourceSerializer";
+import {
+  buildQueue,
+  buildQueueOptions,
+  buildQueueRuntimeInfo,
   InternalQueueOptions,
   QueueDescription,
-  buildQueueOptions,
-  buildQueue,
-  QueueResourceSerializer
+  QueueResourceSerializer,
+  QueueRuntimeInfo
 } from "./serializers/queueResourceSerializer";
 import {
-  InternalTopicOptions,
-  buildTopicOptions,
-  TopicDescription,
-  buildTopic,
-  TopicResourceSerializer
-} from "./serializers/topicResourceSerializer";
-import {
-  InternalSubscriptionOptions,
-  SubscriptionDescription,
-  buildSubscriptionOptions,
-  buildSubscription,
-  SubscriptionResourceSerializer
-} from "./serializers/subscriptionResourceSerializer";
-import {
-  RuleDescription,
   buildRule,
+  RuleDescription,
   RuleResourceSerializer
 } from "./serializers/ruleResourceSerializer";
-import { isJSONLikeObject, isAbsoluteUrl } from "./util/utils";
+import {
+  buildSubscription,
+  buildSubscriptionOptions,
+  buildSubscriptionRuntimeInfo,
+  InternalSubscriptionOptions,
+  SubscriptionDescription,
+  SubscriptionResourceSerializer,
+  SubscriptionRuntimeInfo
+} from "./serializers/subscriptionResourceSerializer";
+import {
+  buildTopic,
+  buildTopicOptions,
+  buildTopicRuntimeInfo,
+  InternalTopicOptions,
+  TopicDescription,
+  TopicResourceSerializer,
+  TopicRuntimeInfo
+} from "./serializers/topicResourceSerializer";
+import { AtomXmlSerializer, executeAtomXmlOperation } from "./util/atomXmlHelper";
+import * as Constants from "./util/constants";
+import { SasServiceClientCredentials } from "./util/sasServiceClientCredentials";
+import { isAbsoluteUrl, isJSONLikeObject } from "./util/utils";
 
 /**
  * Options to use with ServiceBusManagementClient creation
@@ -77,7 +84,41 @@ export interface ListRequestOptions {
   skip?: number;
 }
 /**
- * Represents result of create, get and update operations on queue.
+ * Represents properties of the namespace.
+ */
+export interface NamespaceResponse extends NamespaceProperties {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpOperationResponse;
+}
+/**
+ * Represents properties of the namespace.
+ */
+export type GetNamespaceResponse = NamespaceResponse;
+
+/**
+ * Represents result of create, get, update and delete operations on queue.
+ */
+export interface GetQueueRuntimeInfoResponse extends QueueRuntimeInfo {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpOperationResponse;
+}
+
+/**
+ * Represents result of create, get, update and delete operations on queue.
+ */
+export interface GetQueuesRuntimeInfoResponse extends Array<QueueRuntimeInfo> {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpOperationResponse;
+}
+
+/**
+ * Represents result of create, get, update and delete operations on queue.
  */
 export interface QueueResponse extends QueueDescription {
   /**
@@ -166,6 +207,26 @@ export interface GetTopicsResponse extends Array<TopicDescription> {
 }
 
 /**
+ * Represents result of create, get, update and delete operations on topic.
+ */
+export interface GetTopicRuntimeInfoResponse extends TopicRuntimeInfo {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpOperationResponse;
+}
+
+/**
+ * Represents result of create, get, update and delete operations on topic.
+ */
+export interface GetTopicsRuntimeInfoResponse extends Array<TopicRuntimeInfo> {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpOperationResponse;
+}
+
+/**
  * Represents result of create, get, update and delete operations on subscription.
  */
 export interface SubscriptionResponse extends SubscriptionDescription {
@@ -204,6 +265,26 @@ export interface DeleteSubscriptionResponse {
  * Represents result of list operation on subscriptions.
  */
 export interface GetSubscriptionsResponse extends Array<SubscriptionDescription> {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpOperationResponse;
+}
+
+/**
+ * Represents result of create, get, update and delete operations on topic.
+ */
+export interface GetSubscriptionRuntimeInfoResponse extends SubscriptionRuntimeInfo {
+  /**
+   * The underlying HTTP response.
+   */
+  _response: HttpOperationResponse;
+}
+
+/**
+ * Represents result of create, get, update and delete operations on topic.
+ */
+export interface GetSubscriptionsRuntimeInfoResponse extends Array<SubscriptionRuntimeInfo> {
   /**
    * The underlying HTTP response.
    */
@@ -273,6 +354,7 @@ export class ServiceBusManagementClient extends ServiceClient {
   /**
    * Singleton instances of serializers used across the various operations.
    */
+  private namespaceResourceSerializer: AtomXmlSerializer;
   private queueResourceSerializer: AtomXmlSerializer;
   private topicResourceSerializer: AtomXmlSerializer;
   private subscriptionResourceSerializer: AtomXmlSerializer;
@@ -281,7 +363,7 @@ export class ServiceBusManagementClient extends ServiceClient {
   /**
    * SAS token provider used to generate tokens as required for the various operations.
    */
-  private sasTokenProvider: SharedKeyCredential | TokenCredential;
+  private sasTokenProvider: SharedKeyCredential;
 
   /**
    * Initializes a new instance of the ServiceBusManagementClient class.
@@ -311,10 +393,6 @@ export class ServiceBusManagementClient extends ServiceClient {
     };
 
     super(credentials, serviceClientOptions);
-    this.queueResourceSerializer = new QueueResourceSerializer();
-    this.topicResourceSerializer = new TopicResourceSerializer();
-    this.subscriptionResourceSerializer = new SubscriptionResourceSerializer();
-    this.ruleResourceSerializer = new RuleResourceSerializer();
     this.endpoint = (connectionString.match("Endpoint=.*://(.*)/;") || "")[1];
     this.endpointWithProtocol = connectionStringObj.Endpoint;
 
@@ -322,6 +400,27 @@ export class ServiceBusManagementClient extends ServiceClient {
       connectionStringObj.SharedAccessKeyName,
       connectionStringObj.SharedAccessKey
     );
+
+    this.namespaceResourceSerializer = new NamespaceResourceSerializer();
+    this.queueResourceSerializer = new QueueResourceSerializer();
+    this.topicResourceSerializer = new TopicResourceSerializer();
+    this.subscriptionResourceSerializer = new SubscriptionResourceSerializer();
+    this.ruleResourceSerializer = new RuleResourceSerializer();
+  }
+
+  /**
+   * Returns an object representing the metadata related to a service bus namespace.
+   * @param queueName
+   *
+   */
+  async getNamespaceProperties(): Promise<GetNamespaceResponse> {
+    log.httpAtomXml(`Performing management operation - getNamespaceProperties()`);
+    const response: HttpOperationResponse = await this.getResource(
+      "$namespaceinfo",
+      this.namespaceResourceSerializer
+    );
+
+    return this.buildNamespaceResponse(response);
   }
 
   /**
@@ -378,7 +477,8 @@ export class ServiceBusManagementClient extends ServiceClient {
   }
 
   /**
-   * Returns an object representing the Queue with the given name along with all its properties
+   * Returns an object representing the Queue and its properties.
+   * If you want to get the Queue runtime info like message count details, use `getQueueRuntimeInfo` API.
    * @param queueName
    *
    * Following are errors that can be expected from this operation
@@ -402,7 +502,32 @@ export class ServiceBusManagementClient extends ServiceClient {
   }
 
   /**
-   * Lists existing queues.
+   * Returns an object representing the Queue runtime info like message count details.
+   * @param queueName
+   *
+   * Following are errors that can be expected from this operation
+   * @throws `RestError` with code `UnauthorizedRequestError` when given request fails due to authorization problems,
+   * @throws `RestError` with code `MessageEntityNotFoundError` when requested messaging entity does not exist,
+   * @throws `RestError` with code `InvalidOperationError` when requested operation is invalid and we encounter a 403 HTTP status code,
+   * @throws `RestError` with code `ServerBusyError` when the request fails due to server being busy,
+   * @throws `RestError` with code `ServiceError` when receiving unrecognized HTTP status or for a scenarios such as
+   * bad requests or requests resulting in conflicting operation on the server,
+   * @throws `RestError` with code that is a value from the standard set of HTTP status codes as documented at
+   * https://docs.microsoft.com/en-us/dotnet/api/system.net.httpstatuscode?view=netframework-4.8
+   */
+  async getQueueRuntimeInfo(queueName: string): Promise<GetQueueRuntimeInfoResponse> {
+    log.httpAtomXml(`Performing management operation - getQueue() for "${queueName}"`);
+    const response: HttpOperationResponse = await this.getResource(
+      queueName,
+      this.queueResourceSerializer
+    );
+
+    return this.buildQueueRuntimeInfoResponse(response);
+  }
+
+  /**
+   * Returns a list of objects, each representing a Queue along with its properties.
+   * If you want to get the runtime info of the queues like message count, use `getQueuesRuntimeInfo` API instead.
    * @param listRequestOptions
    *
    * Following are errors that can be expected from this operation
@@ -425,6 +550,34 @@ export class ServiceBusManagementClient extends ServiceClient {
     );
 
     return this.buildListQueuesResponse(response);
+  }
+
+  /**
+   * Returns a list of objects, each representing a Queue's runtime info like message count details.
+   * @param listRequestOptions
+   *
+   * Following are errors that can be expected from this operation
+   * @throws `RestError` with code `UnauthorizedRequestError` when given request fails due to authorization problems,
+   * @throws `RestError` with code `InvalidOperationError` when requested operation is invalid and we encounter a 403 HTTP status code,
+   * @throws `RestError` with code `ServerBusyError` when the request fails due to server being busy,
+   * @throws `RestError` with code `ServiceError` when receiving unrecognized HTTP status or for a scenarios such as
+   * bad requests or requests resulting in conflicting operation on the server,
+   * @throws `RestError` with code that is a value from the standard set of HTTP status codes as documented at
+   * https://docs.microsoft.com/en-us/dotnet/api/system.net.httpstatuscode?view=netframework-4.8
+   */
+  async getQueuesRuntimeInfo(
+    listRequestOptions?: ListRequestOptions
+  ): Promise<GetQueuesRuntimeInfoResponse> {
+    log.httpAtomXml(
+      `Performing management operation - listQueues() with options: ${listRequestOptions}`
+    );
+    const response: HttpOperationResponse = await this.listResources(
+      "$Resources/Queues",
+      listRequestOptions,
+      this.queueResourceSerializer
+    );
+
+    return this.buildListQueuesRuntimeInfoResponse(response);
   }
 
   /**
@@ -496,6 +649,23 @@ export class ServiceBusManagementClient extends ServiceClient {
   }
 
   /**
+   * Checks whether a given queue exists or not.
+   * @param queueName
+   */
+  async queueExists(queueName: string): Promise<boolean> {
+    log.httpAtomXml(`Performing management operation - queueExists() for "${queueName}"`);
+    try {
+      await this.getQueue(queueName);
+    } catch (error) {
+      if (error.code == "MessageEntityNotFoundError") {
+        return false;
+      }
+      throw error;
+    }
+    return true;
+  }
+
+  /**
    * Creates a topic with given name, configured using the given options
    * @param topicName
    *
@@ -549,7 +719,8 @@ export class ServiceBusManagementClient extends ServiceClient {
   }
 
   /**
-   * Returns an object representing the Topic with the given name along with all its properties
+   * Returns an object representing the Topic and its properties.
+   * If you want to get the Topic runtime info like subscription count details, use `getTopicRuntimeInfo` API.
    * @param topicName
    *
    * Following are errors that can be expected from this operation
@@ -573,7 +744,32 @@ export class ServiceBusManagementClient extends ServiceClient {
   }
 
   /**
-   * Lists existing topics.
+   * Returns an object representing the Topic runtime info like subscription count.
+   * @param topicName
+   *
+   * Following are errors that can be expected from this operation
+   * @throws `RestError` with code `UnauthorizedRequestError` when given request fails due to authorization problems,
+   * @throws `RestError` with code `MessageEntityNotFoundError` when requested messaging entity does not exist,
+   * @throws `RestError` with code `InvalidOperationError` when requested operation is invalid and we encounter a 403 HTTP status code,
+   * @throws `RestError` with code `ServerBusyError` when the request fails due to server being busy,
+   * @throws `RestError` with code `ServiceError` when receiving unrecognized HTTP status or for a scenarios such as
+   * bad requests or requests resulting in conflicting operation on the server,
+   * @throws `RestError` with code that is a value from the standard set of HTTP status codes as documented at
+   * https://docs.microsoft.com/en-us/dotnet/api/system.net.httpstatuscode?view=netframework-4.8
+   */
+  async getTopicRuntimeInfo(topicName: string): Promise<GetTopicRuntimeInfoResponse> {
+    log.httpAtomXml(`Performing management operation - getTopicRuntimeInfo() for "${topicName}"`);
+    const response: HttpOperationResponse = await this.getResource(
+      topicName,
+      this.topicResourceSerializer
+    );
+
+    return this.buildTopicRuntimeInfoResponse(response);
+  }
+
+  /**
+   * Returns a list of objects, each representing a Topic along with its properties.
+   * If you want to get the runtime info of the topics like subscription count, use `getTopicsRuntimeInfo` API instead.
    * @param listRequestOptions
    *
    * Following are errors that can be expected from this operation
@@ -596,6 +792,34 @@ export class ServiceBusManagementClient extends ServiceClient {
     );
 
     return this.buildListTopicsResponse(response);
+  }
+
+  /**
+   * Returns a list of objects, each representing a Topic's runtime info like subscription count.
+   * @param listRequestOptions
+   *
+   * Following are errors that can be expected from this operation
+   * @throws `RestError` with code `UnauthorizedRequestError` when given request fails due to authorization problems,
+   * @throws `RestError` with code `InvalidOperationError` when requested operation is invalid and we encounter a 403 HTTP status code,
+   * @throws `RestError` with code `ServerBusyError` when the request fails due to server being busy,
+   * @throws `RestError` with code `ServiceError` when receiving unrecognized HTTP status or for a scenarios such as
+   * bad requests or requests resulting in conflicting operation on the server,
+   * @throws `RestError` with code that is a value from the standard set of HTTP status codes as documented at
+   * https://docs.microsoft.com/en-us/dotnet/api/system.net.httpstatuscode?view=netframework-4.8
+   */
+  async getTopicsRuntimeInfo(
+    listRequestOptions?: ListRequestOptions
+  ): Promise<GetTopicsRuntimeInfoResponse> {
+    log.httpAtomXml(
+      `Performing management operation - listTopics() with options: ${listRequestOptions}`
+    );
+    const response: HttpOperationResponse = await this.listResources(
+      "$Resources/Topics",
+      listRequestOptions,
+      this.topicResourceSerializer
+    );
+
+    return this.buildListTopicsRuntimeInfoResponse(response);
   }
 
   /**
@@ -664,6 +888,23 @@ export class ServiceBusManagementClient extends ServiceClient {
     );
 
     return { _response: response };
+  }
+
+  /**
+   * Checks whether a given topic exists or not.
+   * @param topicName
+   */
+  async topicExists(topicName: string): Promise<boolean> {
+    log.httpAtomXml(`Performing management operation - topicExists() for "${topicName}"`);
+    try {
+      await this.getTopic(topicName);
+    } catch (error) {
+      if (error.code == "MessageEntityNotFoundError") {
+        return false;
+      }
+      throw error;
+    }
+    return true;
   }
 
   /**
@@ -740,7 +981,8 @@ export class ServiceBusManagementClient extends ServiceClient {
   }
 
   /**
-   * Returns an object representing the Subscription with the given name along with all its properties
+   * Returns an object representing the Subscription and its properties.
+   * If you want to get the Subscription runtime info like message count details, use `getSubscriptionRuntimeInfo` API.
    * @param topicName
    * @param subscriptionName
    *
@@ -767,11 +1009,43 @@ export class ServiceBusManagementClient extends ServiceClient {
       this.subscriptionResourceSerializer
     );
 
-    return this.buildSubscriptionResponse(response);
+    return this.buildSubscriptionRuntimeInfoResponse(response);
   }
 
   /**
-   * Lists existing subscriptions.
+   * Returns an object representing the Subscription runtime info like message count details.
+   * @param topicName
+   * @param subscriptionName
+   *
+   * Following are errors that can be expected from this operation
+   * @throws `RestError` with code `UnauthorizedRequestError` when given request fails due to authorization problems,
+   * @throws `RestError` with code `MessageEntityNotFoundError` when requested messaging entity does not exist,
+   * @throws `RestError` with code `InvalidOperationError` when requested operation is invalid and we encounter a 403 HTTP status code,
+   * @throws `RestError` with code `ServerBusyError` when the request fails due to server being busy,
+   * @throws `RestError` with code `ServiceError` when receiving unrecognized HTTP status or for a scenarios such as
+   * bad requests or requests resulting in conflicting operation on the server,
+   * @throws `RestError` with code that is a value from the standard set of HTTP status codes as documented at
+   * https://docs.microsoft.com/en-us/dotnet/api/system.net.httpstatuscode?view=netframework-4.8
+   */
+  async getSubscriptionRuntimeInfo(
+    topicName: string,
+    subscriptionName: string
+  ): Promise<GetSubscriptionRuntimeInfoResponse> {
+    log.httpAtomXml(
+      `Performing management operation - getSubscription() for "${subscriptionName}"`
+    );
+    const fullPath = this.getSubscriptionPath(topicName, subscriptionName);
+    const response: HttpOperationResponse = await this.getResource(
+      fullPath,
+      this.subscriptionResourceSerializer
+    );
+
+    return this.buildSubscriptionRuntimeInfoResponse(response);
+  }
+
+  /**
+   * Returns a list of objects, each representing a Subscription along with its properties.
+   * If you want to get the runtime info of the subscriptions like message count, use `getSubscriptionsRuntimeInfo` API instead.
    * @param topicName
    * @param listRequestOptions
    *
@@ -798,6 +1072,36 @@ export class ServiceBusManagementClient extends ServiceClient {
     );
 
     return this.buildListSubscriptionsResponse(response);
+  }
+
+  /**
+   * Returns a list of objects, each representing a Subscription's runtime info like message count details.
+   * @param topicName
+   * @param listRequestOptions
+   *
+   * Following are errors that can be expected from this operation
+   * @throws `RestError` with code `UnauthorizedRequestError` when given request fails due to authorization problems,
+   * @throws `RestError` with code `InvalidOperationError` when requested operation is invalid and we encounter a 403 HTTP status code,
+   * @throws `RestError` with code `ServerBusyError` when the request fails due to server being busy,
+   * @throws `RestError` with code `ServiceError` when receiving unrecognized HTTP status or for a scenarios such as
+   * bad requests or requests resulting in conflicting operation on the server,
+   * @throws `RestError` with code that is a value from the standard set of HTTP status codes as documented at
+   * https://docs.microsoft.com/en-us/dotnet/api/system.net.httpstatuscode?view=netframework-4.8
+   */
+  async getSubscriptionsRuntimeInfo(
+    topicName: string,
+    listRequestOptions?: ListRequestOptions
+  ): Promise<GetSubscriptionsRuntimeInfoResponse> {
+    log.httpAtomXml(
+      `Performing management operation - listSubscriptions() with options: ${listRequestOptions}`
+    );
+    const response: HttpOperationResponse = await this.listResources(
+      topicName + "/Subscriptions/",
+      listRequestOptions,
+      this.subscriptionResourceSerializer
+    );
+
+    return this.buildListSubscriptionsRuntimeInfoResponse(response);
   }
 
   /**
@@ -888,6 +1192,27 @@ export class ServiceBusManagementClient extends ServiceClient {
     );
 
     return { _response: response };
+  }
+
+  /**
+   * Checks whether a given subscription exists in the topic or not.
+   * @param topicName
+   * @param subscriptionName
+   *
+   */
+  async subscriptionExists(topicName: string, subscriptionName: string): Promise<boolean> {
+    log.httpAtomXml(
+      `Performing management operation - subscriptionExists() for "${topicName}" and "${subscriptionName}"`
+    );
+    try {
+      await this.getSubscription(topicName, subscriptionName);
+    } catch (error) {
+      if (error.code == "MessageEntityNotFoundError") {
+        return false;
+      }
+      throw error;
+    }
+    return true;
   }
 
   /**
@@ -1212,6 +1537,25 @@ export class ServiceBusManagementClient extends ServiceClient {
     return topicName + "/Subscriptions/" + subscriptionName + "/Rules/" + ruleName;
   }
 
+  private buildNamespaceResponse(response: HttpOperationResponse): NamespaceResponse {
+    try {
+      const namespace = buildNamespace(response.parsedBody);
+      const namespaceResponse: NamespaceResponse = Object.assign(namespace || {}, {
+        _response: response
+      });
+      return namespaceResponse;
+    } catch (err) {
+      log.warning("Failure parsing response from service - %0 ", err);
+      throw new RestError(
+        `Error occurred while parsing the response body - cannot form a namespace object using the response from the service.`,
+        RestError.PARSE_ERROR,
+        response.status,
+        stripRequest(response.request),
+        stripResponse(response)
+      );
+    }
+  }
+
   private buildListQueuesResponse(response: HttpOperationResponse): GetQueuesResponse {
     try {
       const queues: QueueDescription[] = [];
@@ -1241,10 +1585,62 @@ export class ServiceBusManagementClient extends ServiceClient {
     }
   }
 
+  private buildListQueuesRuntimeInfoResponse(
+    response: HttpOperationResponse
+  ): GetQueuesRuntimeInfoResponse {
+    try {
+      const queues: QueueRuntimeInfo[] = [];
+      if (!Array.isArray(response.parsedBody)) {
+        throw new TypeError(`${response.parsedBody} was expected to be of type Array`);
+      }
+      const rawQueueArray: any = response.parsedBody;
+      for (let i = 0; i < rawQueueArray.length; i++) {
+        const queue = buildQueueRuntimeInfo(rawQueueArray[i]);
+        if (queue) {
+          queues.push(queue);
+        }
+      }
+      const listQueuesResponse: GetQueuesRuntimeInfoResponse = Object.assign(queues, {
+        _response: response
+      });
+      return listQueuesResponse;
+    } catch (err) {
+      log.warning("Failure parsing response from service - %0 ", err);
+      throw new RestError(
+        `Error occurred while parsing the response body - cannot form a list of queues using the response from the service.`,
+        RestError.PARSE_ERROR,
+        response.status,
+        stripRequest(response.request),
+        stripResponse(response)
+      );
+    }
+  }
+
   private buildQueueResponse(response: HttpOperationResponse): QueueResponse {
     try {
       const queue = buildQueue(response.parsedBody);
       const queueResponse: QueueResponse = Object.assign(queue || {}, {
+        _response: response
+      });
+      return queueResponse;
+    } catch (err) {
+      log.warning("Failure parsing response from service - %0 ", err);
+      throw new RestError(
+        `Error occurred while parsing the response body - cannot form a queue object using the response from the service.`,
+        RestError.PARSE_ERROR,
+        response.status,
+        stripRequest(response.request),
+        stripResponse(response)
+      );
+    }
+  }
+
+  private buildQueueRuntimeInfoResponse(
+    response: HttpOperationResponse
+  ): GetQueueRuntimeInfoResponse {
+    try {
+      const queue = buildQueueRuntimeInfo(response.parsedBody);
+      const queueResponse: GetQueueRuntimeInfoResponse = Object.assign(queue || {}, {
         _response: response
       });
       return queueResponse;
@@ -1288,10 +1684,62 @@ export class ServiceBusManagementClient extends ServiceClient {
       );
     }
   }
+
+  private buildListTopicsRuntimeInfoResponse(
+    response: HttpOperationResponse
+  ): GetTopicsRuntimeInfoResponse {
+    try {
+      const topics: TopicRuntimeInfo[] = [];
+      if (!Array.isArray(response.parsedBody)) {
+        throw new TypeError(`${response.parsedBody} was expected to be of type Array`);
+      }
+      const rawTopicArray: any = response.parsedBody;
+      for (let i = 0; i < rawTopicArray.length; i++) {
+        const topic = buildTopicRuntimeInfo(rawTopicArray[i]);
+        if (topic) {
+          topics.push(topic);
+        }
+      }
+      const listTopicsResponse: GetTopicsRuntimeInfoResponse = Object.assign(topics, {
+        _response: response
+      });
+      return listTopicsResponse;
+    } catch (err) {
+      log.warning("Failure parsing response from service - %0 ", err);
+      throw new RestError(
+        `Error occurred while parsing the response body - cannot form a list of topics using the response from the service.`,
+        RestError.PARSE_ERROR,
+        response.status,
+        stripRequest(response.request),
+        stripResponse(response)
+      );
+    }
+  }
   private buildTopicResponse(response: HttpOperationResponse): TopicResponse {
     try {
       const topic = buildTopic(response.parsedBody);
       const topicResponse: TopicResponse = Object.assign(topic || {}, {
+        _response: response
+      });
+      return topicResponse;
+    } catch (err) {
+      log.warning("Failure parsing response from service - %0 ", err);
+      throw new RestError(
+        `Error occurred while parsing the response body - cannot form a topic object using the response from the service.`,
+        RestError.PARSE_ERROR,
+        response.status,
+        stripRequest(response.request),
+        stripResponse(response)
+      );
+    }
+  }
+
+  private buildTopicRuntimeInfoResponse(
+    response: HttpOperationResponse
+  ): GetTopicRuntimeInfoResponse {
+    try {
+      const topic = buildTopicRuntimeInfo(response.parsedBody);
+      const topicResponse: GetTopicRuntimeInfoResponse = Object.assign(topic || {}, {
         _response: response
       });
       return topicResponse;
@@ -1338,12 +1786,70 @@ export class ServiceBusManagementClient extends ServiceClient {
     }
   }
 
+  private buildListSubscriptionsRuntimeInfoResponse(
+    response: HttpOperationResponse
+  ): GetSubscriptionsRuntimeInfoResponse {
+    try {
+      const subscriptions: SubscriptionRuntimeInfo[] = [];
+      if (!Array.isArray(response.parsedBody)) {
+        throw new TypeError(`${response.parsedBody} was expected to be of type Array`);
+      }
+      const rawSubscriptionArray: any = response.parsedBody;
+      for (let i = 0; i < rawSubscriptionArray.length; i++) {
+        const subscription = buildSubscriptionRuntimeInfo(rawSubscriptionArray[i]);
+        if (subscription) {
+          subscriptions.push(subscription);
+        }
+      }
+      const listSubscriptionsResponse: GetSubscriptionsRuntimeInfoResponse = Object.assign(
+        subscriptions,
+        {
+          _response: response
+        }
+      );
+      return listSubscriptionsResponse;
+    } catch (err) {
+      log.warning("Failure parsing response from service - %0 ", err);
+      throw new RestError(
+        `Error occurred while parsing the response body - cannot form a list of subscriptions using the response from the service.`,
+        RestError.PARSE_ERROR,
+        response.status,
+        stripRequest(response.request),
+        stripResponse(response)
+      );
+    }
+  }
+
   private buildSubscriptionResponse(response: HttpOperationResponse): SubscriptionResponse {
     try {
       const subscription = buildSubscription(response.parsedBody);
       const subscriptionResponse: SubscriptionResponse = Object.assign(subscription || {}, {
         _response: response
       });
+      return subscriptionResponse;
+    } catch (err) {
+      log.warning("Failure parsing response from service - %0 ", err);
+      throw new RestError(
+        `Error occurred while parsing the response body - cannot form a subscription object using the response from the service.`,
+        RestError.PARSE_ERROR,
+        response.status,
+        stripRequest(response.request),
+        stripResponse(response)
+      );
+    }
+  }
+
+  private buildSubscriptionRuntimeInfoResponse(
+    response: HttpOperationResponse
+  ): GetSubscriptionRuntimeInfoResponse {
+    try {
+      const subscription = buildSubscriptionRuntimeInfo(response.parsedBody);
+      const subscriptionResponse: GetSubscriptionRuntimeInfoResponse = Object.assign(
+        subscription || {},
+        {
+          _response: response
+        }
+      );
       return subscriptionResponse;
     } catch (err) {
       log.warning("Failure parsing response from service - %0 ", err);