+++ sdk/servicebus/service-bus/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Improves the performance of the `ServiceBusMessageBatch.tryAdd` method.
   [PR 8772](https://github.com/Azure/azure-sdk-for-js/pull/8772)
+- Added management api features which allows CRUD operations on the entities of a namespace.
+  [PR 9116](https://github.com/Azure/azure-sdk-for-js/pull/9116)
+  [PR 9221](https://github.com/Azure/azure-sdk-for-js/pull/9221)
 
 ## 7.0.0-preview.2 (2020-05-05)